+++ src/test/java/kitchenpos/apply/menu/tobe/application/MenuServiceTest.java
@@ -8,7 +8,6 @@ import kitchenpos.apply.menus.tobe.application.MenuService;
 import kitchenpos.apply.menus.tobe.domain.MenuGroupRepository;
 import kitchenpos.apply.menus.tobe.domain.MenuPriceChecker;
 import kitchenpos.apply.menus.tobe.domain.MenuRepository;
-import kitchenpos.apply.menus.tobe.domain.PurgomalumClient;
 import kitchenpos.apply.menus.tobe.ui.MenuProductRequest;
 import kitchenpos.apply.menus.tobe.ui.MenuRequest;
 import kitchenpos.apply.menus.tobe.ui.MenuResponse;
@@ -16,6 +15,7 @@ import kitchenpos.apply.products.tobe.application.ProductValidator;
 import kitchenpos.apply.products.tobe.domain.InMemoryProductRepository;
 import kitchenpos.apply.products.tobe.domain.Product;
 import kitchenpos.apply.products.tobe.domain.ProductRepository;
+import kitchenpos.support.domain.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
